fix(epistemic): five interlocking architectural fixes for graph control + posterior accounting
Diagnosed against smoke_v6_nano (case 847's 513-iteration runaway, 957's
abandoned-but-real-evidence collapse). Each fix is general — none is
case-specific. Five cohorts of changes, each addressing one issue from
ARCHITECTURE_ANALYSIS.md.

A5 — sibling-resolution did_work flag

  ResolveUncertaintyOperation does sibling-grouping: one LLM call resolves
  N near-duplicate uncertainties at once. The remaining N-1 graph-loop
  iterations short-circuit with "Already resolved" but were still flagging
  affected claims for re-scrutiny. Added did_work to OperationResult; the
  Resolve op sets it False on no-op paths; the graph node only marks for
  re-scrutiny on real work.

A1 — scrutiny idempotence via input fingerprint

  ScrutiniseClaimOperation was stateless: every call minted fresh
  Uncertainty entities for the same issues. Added scrutiny_fingerprint to
  Claim (SHA-256 over claim.statement, scope, sorted eligible
  evidence_ids). The operation short-circuits with did_work=False when the
  fingerprint matches the previous verdict's inputs; the Scrutinize graph
  node clears the fingerprint when honoring claims_needing_rescrutiny so
  intentional re-passes still run. Persisted via the standard
  _extra_metadata / _from_metadata pattern.

A2 — convergence verdict becomes a real termination edge

  AssessConvergenceOperation produced a CONVERGENT verdict that no graph
  node ever read. Added convergence_verdict field to Claim; the operation
  now writes it before saving. RunVerification fast-paths to
  IntegrateEvidence when any active SUPPORTED claim is CONVERGENT and no
  blocking uncertainties remain — otherwise existing ResolveUncertainties
  path. Conservative guard: convergence speeds up only when it is safe to.

A3 — soft-promote path preserves abandoned-claim signal

  Refute-promotion's "n_con >= 2 * max(1, n_sup)" threshold and abandonment
  were the only two outcomes when promote_claim declined. Abandoned claims
  are filtered out of compute_posterior, erasing their evidence record.
  Added SoftPromoteOperation: when refute declines but linked evidence
  carries directional judgments (n_sup + n_con > 0), promote to SUPPORTED
  with integrated_assessment="insufficient" and confidence
  min(0.5, max(n_sup, n_con) / (n_sup + n_con + 1)). AbandonOrDemote tries
  refute, then soft-promote, then abandons only when there is genuinely no
  signal. The give-up paths are now symmetric.

A4 — posterior weights by cluster size; cap moves to LLM consumers

  EVIDENCE_TOP_K=5 conflated two unrelated decisions: how many clusters
  the LLM panel inspects (cost) and how many contribute to the posterior
  (signal). Result: ~95% of clustered evidence was tagged "deferred" and
  invisible to the posterior. Replaced with LLM_PANEL_CAP=10 living in
  the consumers (assess_convergence per-rep classify + pairwise
  independence; validate_deductively; verify_computationally;
  integrate_evidence; synthesize_report; scrutinise's evidence-summary
  prompt). select_top_k_evidence now picks representatives for every
  cluster — no more "deferred" status. compute_posterior weights each
  representative by 1 + log(corroboration_count) so a cluster of 50
  contributes ~4.9, a singleton contributes 1.0. supporting_count and
  contradicting_count are now floats (semantics: weighted count).

Tests updated: drift checksum bumped; the two old top-K tests rewritten
as "every cluster gets a representative, none deferred"; refute test
split into no-signal (abandons) and directional-signal (soft-promoted);
report-fields test expects float types. 1588 passed, 1 skipped, 6
deselected.

+++ src/andamentum/epistemic/report_data.py
@@ -103,8 +103,8 @@ class ConfidenceScores:
     """Answer-level confidence scores for report rendering."""
 
     posterior: Optional[float] = None
-    posterior_supporting: int = 0
-    posterior_contradicting: int = 0
+    posterior_supporting: float = 0.0
+    posterior_contradicting: float = 0.0
     posterior_question_type: Optional[str] = None
     terminal_state: str = "completed"
